# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bot-Mox API v1
  version: 1.0.0
  description: Canonical API surface for phased backend/frontend alignment.
servers:
  - url: http://localhost:3002
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ResourceMutationRequest:
      description: |
        Kind-specific mutation payload. Server validates by `{kind}`.
        - `licenses`: `LicenseMutationRequest`
        - `proxies`: `ProxyMutationRequest`
        - `subscriptions`: `SubscriptionMutationRequest`
      oneOf:
        - $ref: '#/components/schemas/LicenseMutationRequest'
        - $ref: '#/components/schemas/ProxyMutationRequest'
        - $ref: '#/components/schemas/SubscriptionMutationRequest'
...
    LicenseMutationRequest:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [active, expired, revoked]
        bot_ids:
          type: array
          items:
            type: string
        expires_at:
          type: integer
      additionalProperties: true
    ProxyMutationRequest:
      type: object
      properties:
        id:
          type: string
        ip:
          type: string
        port:
          type: integer
        type:
          type: string
          enum: [http, socks5]
        status:
          type: string
          enum: [active, expired, banned]
        bot_id:
          type: string
          nullable: true
      additionalProperties: true
    SubscriptionMutationRequest:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [wow, bot, proxy, vpn, other]
        status:
          type: string
          enum: [active, cancelled, expired, expiring_soon]
        bot_id:
          type: string
        project_id:
          type: string
          enum: [wow_tbc, wow_midnight]
        expires_at:
          type: integer
      additionalProperties: true
...
    LicenseLeaseRequest:
      type: object
      properties:
        vm_uuid:
          type: string
        user_id:
          type: string
        agent_id:
          type: string
        runner_id:
          type: string
        module:
          type: string
        version:
          type: string
      required: [vm_uuid, agent_id, runner_id, module]
    LicenseHeartbeatRequest:
      type: object
      properties:
        lease_id:
          type: string
      required: [lease_id]
    LicenseRevokeRequest:
      type: object
      properties:
        lease_id:
          type: string
        reason:
          type: string
      required: [lease_id]
...
paths:
...
  /api/v1/resources/{kind}:
    get:
      summary: List resources by kind
      parameters:
        - in: path
          name: kind
          required: true
          schema:
            type: string
            enum: [licenses, proxies, subscriptions]
      responses:
        '200':
          description: Resource list
    post:
      summary: Create resource by kind
      parameters:
        - in: path
          name: kind
          required: true
          schema:
            type: string
            enum: [licenses, proxies, subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceMutationRequest'
      responses:
        '201':
          description: Created resource
  /api/v1/resources/{kind}/{id}:
    get:
      summary: Get resource by id
      parameters:
        - in: path
          name: kind
          required: true
          schema:
            type: string
            enum: [licenses, proxies, subscriptions]
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Resource entity
        '404':
          description: Not found
    patch:
      summary: Update resource by id
      parameters:
        - in: path
          name: kind
          required: true
          schema:
            type: string
            enum: [licenses, proxies, subscriptions]
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceMutationRequest'
      responses:
        '200':
          description: Updated resource
    delete:
      summary: Delete resource by id
      parameters:
        - in: path
          name: kind
          required: true
          schema:
            type: string
            enum: [licenses, proxies, subscriptions]
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deletion result
...
  /api/v1/license/lease:
    post:
      summary: Issue short-lived execution lease for runner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LicenseLeaseRequest'
      responses:
        '201':
          description: Lease issued
        '403':
          description: License, entitlement or VM ownership check failed
  /api/v1/license/heartbeat:
    post:
      summary: Update lease heartbeat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LicenseHeartbeatRequest'
      responses:
        '200':
          description: Heartbeat accepted
        '409':
          description: Lease inactive or expired
  /api/v1/license/revoke:
    post:
      summary: Revoke existing execution lease
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LicenseRevokeRequest'
      responses:
        '200':
          description: Lease revoked
...
  /api/v1/secrets:
    post:
      summary: Store new ciphertext secret (server-blind)
      description: Backend stores only ciphertext and metadata. Never returns plaintext.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                ciphertext:
                  type: string
                alg:
                  type: string
                  default: AES-256-GCM
                key_id:
                  type: string
                nonce:
                  type: string
                aad_meta:
                  type: object
                  additionalProperties: true
              required: [label, ciphertext, key_id, nonce]
      responses:
        '201':
          description: Secret metadata (no ciphertext in response)
  /api/v1/secrets/{id}/meta:
    get:
      summary: Get secret metadata (no ciphertext)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Secret metadata
        '404':
          description: Not found
  /api/v1/secrets/{id}/rotate:
    post:
      summary: Rotate secret ciphertext material
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ciphertext:
                  type: string
                alg:
                  type: string
                key_id:
                  type: string
                nonce:
                  type: string
                aad_meta:
                  type: object
              required: [ciphertext, key_id, nonce]
      responses:
        '200':
          description: Rotated secret metadata
  /api/v1/secrets/bindings:
    post:
      summary: Bind secret to scope (bot/vm/agent/tenant)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scope_type:
                  type: string
                  enum: [bot, vm, agent, tenant]
                scope_id:
                  type: string
                secret_ref:
                  type: string
                  format: uuid
                field_name:
                  type: string
              required: [scope_type, scope_id, secret_ref, field_name]
      responses:
        '201':
          description: Binding created
    get:
      summary: List secret bindings for scope
      parameters:
        - in: query
          name: scope_type
          schema:
            type: string
        - in: query
          name: scope_id
          schema:
            type: string
      responses:
        '200':
          description: Binding list
  # --- VM Ops (agent-mediated) ---
...